refactor: user-season
