--- src/components/user-season-header/index.ts
@@ -1,1 +1,0 @@
-export { UserSeasonHeader } from "./user-season-header";
