--- src/components/user-season-history/index.ts
@@ -1,1 +1,0 @@
-export { UserSeasonHistory } from "./user-season-history";
