+++ src/features/user-season/types/season-types.ts
@@ -1,0 +1,20 @@
+import type { Season } from "@/lib/services/seasons";
+
+export interface UserSeasonHistoryItem {
+  season: Season;
+  userLevel: {
+    level: number;
+    xp: number;
+    updated_at: string;
+  } | null;
+}
+
+export interface UserSeasonHeaderProps {
+  season: Season;
+  userId: string;
+}
+
+export interface UserSeasonHistoryProps {
+  userId: string;
+  seasonHistory: UserSeasonHistoryItem[];
+}
